fixup from rebase; test_mutable -> test/mutable/test_*
